Move AnonymousObject from commons to client/server
